+++ crypto/benches/hashing.rs
@@ -4,12 +4,13 @@ use noah_algebra::prelude::{test_rng, Scalar};
 use noah_algebra::secp256k1::{SECP256K1Fq, SECP256K1G1};
 use noah_crypto::hashing_to_the_curve::ed25519_elligator::Ed25519ElligatorParameters;
 use noah_crypto::hashing_to_the_curve::ed25519_sswu_wb::Ed25519SSWUParameters;
-use noah_crypto::hashing_to_the_curve::ed25519_sw::Ed25519SWParameters;
-use noah_crypto::hashing_to_the_curve::secp256k1_sswu_wb::Secp256k1SSWUParameters;
+use noah_crypto::hashing_to_the_curve::models::elligator::Elligator;
+use noah_crypto::hashing_to_the_curve::models::sswu::SimplifiedSWUMap;
+use noah_crypto::hashing_to_the_curve::models::sw::SWMap;
 use noah_crypto::hashing_to_the_curve::secp256k1_sw::Secp256k1SWParameters;
-use noah_crypto::hashing_to_the_curve::traits::{
-    Elligator, HashingToCurve, SWMap, SimplifiedSWUMap,
-};
+use noah_crypto::hashing_to_the_curve::sswu::Secp256k1SSWUParameters;
+use noah_crypto::hashing_to_the_curve::sw::Ed25519SWParameters;
+use noah_crypto::hashing_to_the_curve::traits::HashingToCurve;
 
 fn bench_ed25519_elligator(c: &mut Criterion) {
     let mut single_group = c.benchmark_group("ed25519_elligator");
