+++ crypto/src/hashing_to_the_curve/ed25519/mod.rs
@@ -1,0 +1,8 @@
+/// Elligator for ed25519.
+pub mod elligator;
+
+/// Simplified SWU map for ed25519.
+pub mod sswu;
+
+/// Shallue-van de Woestijne map for ed25519.
+pub mod sw;
