+++ crypto/src/hashing_to_the_curve/mod.rs
@@ -1,17 +1,11 @@
-/// Module for the secp256k1 simplified SWU map (WB map).
-pub mod secp256k1_sswu_wb;
-
-/// Module for the secp256k1 SW map.
-pub mod secp256k1_sw;
-
-/// Module for the ed25519 SW map.
-pub mod ed25519_sw;
-
-/// Module for the ed25519 simplified SWU map (WB map).
-pub mod ed25519_sswu_wb;
-
-/// Module for the ed25519 elligator.
-pub mod ed25519_elligator;
-
 /// Module for the traits of maps.
 pub mod traits;
+
+/// Collections for models.
+pub mod models;
+
+/// Collections of implementations for ed25519.
+pub mod ed25519;
+
+/// Collections of implementations for secp256k1.
+pub mod secp256k1;
