+++ crypto/src/hashing_to_the_curve/models/mod.rs
@@ -1,0 +1,8 @@
+/// Module for the Shallue-van de Woestijne map.
+pub mod sw;
+
+/// Module for the simplified SWU map.
+pub mod sswu;
+
+/// Module for elligator.
+pub mod elligator;
