+++ crypto/src/hashing_to_the_curve/secp256k1/mod.rs
@@ -1,0 +1,5 @@
+/// Simplified SWU map for secp256k1.
+pub mod sswu;
+
+/// Shallue-van de Woestijne map for secp256k1.
+pub mod sw;
