+++ crypto/src/hashing_to_the_curve/secp256k1/sswu.rs
@@ -1,5 +1,5 @@
 use crate::errors::Result;
-use crate::hashing_to_the_curve::traits::SimplifiedSWUParameters;
+use crate::hashing_to_the_curve::models::sswu::SimplifiedSWUParameters;
 use noah_algebra::secp256k1::SECP256K1G1;
 use noah_algebra::{new_secp256k1_fq, prelude::*, secp256k1::SECP256K1Fq};
 
@@ -52,8 +52,9 @@ impl SimplifiedSWUParameters<SECP256K1G1> for Secp256k1SSWUParameters {
 
 #[cfg(test)]
 mod tests {
-    use crate::hashing_to_the_curve::secp256k1_sswu_wb::Secp256k1SSWUParameters;
-    use crate::hashing_to_the_curve::traits::{HashingToCurve, SimplifiedSWUMap};
+    use crate::hashing_to_the_curve::models::sswu::SimplifiedSWUMap;
+    use crate::hashing_to_the_curve::secp256k1::sswu::Secp256k1SSWUParameters;
+    use crate::hashing_to_the_curve::traits::HashingToCurve;
     use noah_algebra::new_secp256k1_fq;
     use noah_algebra::prelude::{test_rng, Scalar};
     use noah_algebra::secp256k1::{SECP256K1Fq, SECP256K1G1};
